Move Table Header to another component
